Run maze runner tests against Laravel latest

+++ features/support/env.rb
@@ -7,6 +7,8 @@ VENDORED_LIB = FIXTURE_DIR + '/bugsnag-laravel.zip'
 
 `composer archive -f zip --dir=#{File.dirname(VENDORED_LIB)} --file=#{File.basename(VENDORED_LIB, '.zip')}`
 Dir.glob(FIXTURE_DIR + '/laravel*').each do |directory|
+  next if directory.end_with?('laravel-latest')
+
   FileUtils.cp(VENDORED_LIB, directory + '/bugsnag-laravel.zip')
   # Remove any locally installed composer deps
   FileUtils.rm_rf(directory + '/vendor')
@@ -18,6 +20,8 @@ File.open('composer.json', 'r') do |source|
   parsed_composer = JSON.parse source.read
   requirements = parsed_composer["require"]
   Dir.glob(FIXTURE_DIR + '/laravel*').each do |directory|
+    next if directory.end_with?('laravel-latest')
+
     File.open(directory + '/composer.json.template', 'r') do |template|
       parsed_template = JSON.parse template.read
       parsed_template["repositories"][0]["package"]["require"] = requirements
